+++ CHANGELOG.md
@@ -72,6 +72,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 * Glob patterns now support `{foo,bar}` syntax. There may be subtle behavior
   changes as we use the [globset](https://crates.io/crates/globset) library now.
 
+* The new `bisect(x)` revset function can help bisect a range of commits to
+  find when a bug was introduced.
+
 ### Fixed bugs
 
 ### Packaging changes